requirements.txt
# 电商选品上架 Agent - 依赖

# 核心依赖
pyyaml>=6.0
requests>=2.28.0
aiohttp>=3.8.0

# 图片处理
Pillow>=9.0.0

# 浏览器自动化（可选，OpenClaw 已内置 browser 工具）
# playwright>=1.40.0

# 数据处理
pandas>=1.5.0

# 日志
python-json-logger>=2.0.0
